# encoding/json: do not fill d.errorContext until deemed necessary