This will increase performance, as long as most of the unmarshalled JSON
does not contain errors.

The performance difference in the included benchmarks does not look
promising, but I claim this is only because these benchmarks include
many tests that produce errors:

	name                  old time/op    new time/op    delta
	CodeUnmarshal-2          165ms ± 1%     165ms ± 1%     ~     (p=0.481 n=10+10)
	CodeUnmarshalReuse-2     137ms ± 0%     137ms ± 1%   -0.62%  (p=0.013 n=10+9)
	UnmarshalString-2       1.07µs ± 1%    1.16µs ± 0%   +7.70%  (p=0.000 n=10+10)
	UnmarshalFloat64-2      1.01µs ± 3%    1.08µs ± 1%   +7.39%  (p=0.000 n=9+10)
	UnmarshalInt64-2         845ns ± 0%     978ns ± 0%  +15.82%  (p=0.000 n=9+9)

	name                  old speed      new speed      delta
	CodeUnmarshal-2       11.8MB/s ± 1%  11.8MB/s ± 1%     ~     (p=0.467 n=10+10)
	CodeUnmarshalReuse-2  14.1MB/s ± 0%  14.2MB/s ± 1%   +0.62%  (p=0.014 n=10+9)

	name                  old alloc/op   new alloc/op   delta
	CodeUnmarshal-2         3.28MB ± 0%    3.27MB ± 0%   -0.03%  (p=0.000 n=10+9)
	CodeUnmarshalReuse-2    2.19MB ± 0%    2.19MB ± 0%   -0.05%  (p=0.000 n=9+9)
	UnmarshalString-2         192B ± 0%      208B ± 0%   +8.33%  (p=0.000 n=10+10)
	UnmarshalFloat64-2        180B ± 0%      196B ± 0%   +8.89%  (p=0.000 n=10+10)
	UnmarshalInt64-2          176B ± 0%      192B ± 0%   +9.09%  (p=0.000 n=10+10)

	name                  old allocs/op  new allocs/op  delta
	CodeUnmarshal-2          92.7k ± 0%     92.7k ± 0%   -0.01%  (p=0.000 n=10+10)
	CodeUnmarshalReuse-2     80.4k ± 0%     80.3k ± 0%   -0.01%  (p=0.000 n=10+9)
	UnmarshalString-2         2.00 ± 0%      2.00 ± 0%     ~     (all equal)
	UnmarshalFloat64-2        2.00 ± 0%      2.00 ± 0%     ~     (all equal)
	UnmarshalInt64-2          1.00 ± 0%      1.00 ± 0%     ~     (all equal)

When unmarshaling only valid JSON the results will look more similar to
these [1]:

	name                      old time/op    new time/op    delta
	UnmarshalSmall-2            22.5µs ± 1%    20.9µs ± 2%   -7.09%  (p=0.000 n=10+10)
	UnmarshalSmallOnlyId-2      14.4µs ± 2%    13.7µs ± 1%   -4.41%  (p=0.000 n=10+10)
	UnmarshalMedium-2            202µs ± 2%     195µs ± 1%   -3.47%  (p=0.000 n=10+10)
	UnmarshalMediumOnlyIds-2     131µs ± 1%     127µs ± 1%   -2.69%  (p=0.000 n=10+10)
	UnmarshalLarge-2            2.72ms ± 1%    2.57ms ± 1%   -5.49%  (p=0.000 n=8+10)
	UnmarshalLargeOnlyIds-2     1.62ms ± 1%    1.57ms ± 1%   -2.63%  (p=0.000 n=10+10)

	name                      old alloc/op   new alloc/op   delta
	UnmarshalSmall-2              288B ± 0%      288B ± 0%     ~     (all equal)
	UnmarshalSmallOnlyId-2        248B ± 0%      248B ± 0%     ~     (all equal)
	UnmarshalMedium-2           1.46kB ± 0%    1.36kB ± 0%   -6.59%  (p=0.000 n=10+10)
	UnmarshalMediumOnlyIds-2      392B ± 0%      360B ± 0%   -8.16%  (p=0.000 n=10+10)
	UnmarshalLarge-2            13.4kB ± 0%    13.3kB ± 0%   -0.73%  (p=0.000 n=10+8)
	UnmarshalLargeOnlyIds-2       402B ± 0%      370B ± 0%   -7.96%  (p=0.000 n=10+10)

	name                      old allocs/op  new allocs/op  delta
	UnmarshalSmall-2              8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.000 n=10+10)
	UnmarshalSmallOnlyId-2        4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=10+10)
	UnmarshalMedium-2             49.0 ± 0%      46.0 ± 0%   -6.12%  (p=0.000 n=10+10)
	UnmarshalMediumOnlyIds-2      8.00 ± 0%      6.00 ± 0%  -25.00%  (p=0.000 n=10+10)
	UnmarshalLarge-2               536 ± 0%       533 ± 0%   -0.56%  (p=0.000 n=10+10)
	UnmarshalLargeOnlyIds-2       35.0 ± 0%      33.0 ± 0%   -5.71%  (p=0.000 n=10+10)

[1] https://gist.githubusercontent.com/codesoap/f4fd126d0eb4d534f74f3c9f12ed001b/raw/9c8b3ec5c4b678447b2371430b56c50374bb085c/bench_test.go